+++ animalSimulation/src/animalsimulation/model/knowledge/ObstacleKnowledge.java
@@ -2,7 +2,7 @@ package animalsimulation.model.knowledge;
 
 public class ObstacleKnowledge extends LocationKnowledge{
     
-    public ObstacleKnowledge(int[] argHiveLocation, int[] argTargetLocation) {
+    public ObstacleKnowledge(double[] argHiveLocation, double[] argTargetLocation) {
         super(argHiveLocation, argTargetLocation);
     }
 }
